pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "NREL-emerge"
version = "2.0.0"
description = "Emerging technologies Management and Risk evaluation on distributions Grid Evolution"
readme = "README.md"
license = ""
requires-python = ">=3.11"
authors = [
    { name = "Kapil Duwadi", email = "kapil.duwadi@nrel.gov" },
]
keywords = [
    "DER",
    "Distribution",
    "System",
    "analysis",
    "impact",
    "management",
    "risk",
    "technologies",
]
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "click",
    "json5",
    "networkx",
    "OpenDSSDirect.py==0.8.4",
    "pandas",
    "polars",
    "pyarrow",
    "pydantic",
    "pyyaml",
    "sqlmodel",
    "psutil",
    "loguru"
]

[project.optional-dependencies]
dev = ["pre-commit", "pytest", "pytest-cov", "ruff"]
doc = [
  "sphinx",
  "pydata-sphinx-theme",
  "myst-parser",
  "autodoc_pydantic",
  "sphinxcontrib-mermaid",
]

[project.scripts]
emerge = "emerge.cli.cli:cli"

[project.urls]
Documentation = "https://github.com/NREL/EMeRGE#readme"
Issues = "https://github.com/NREL/EMeRGE/issues"
Source = "https://github.com/NREL/EMeRGE/grid-data-models"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [".git", ".ruff_cache", ".venv", "_build", "build", "dist", "venv"]
ignore-init-module-imports = true
line-length = 99
indent-width = 4
target-version = "py311"


[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = [
  "C901", # McCabe complexity
  "E4",   # Subset of pycodestyle (E)
  "E7",
  "E9",
  "F",    # Pyflakes
  "W",    # pycodestyle warnings
]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"**/{tests,docs,tools}/*" = ["E402"]


[tool.hatch.build.targets.sdist]
include = [
    "/emerge",
]

[tool.hatch.build.targets.wheel]
packages = [
    "emerge"
]
